requirements.txt
streamlit==1.37.1
requests==2.32.3
python-dateutil==2.9.0.post0
pydantic==2.8.2
openai==1.43.0
python-dotenv
google-generativeai